pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "argos"
version = "0.4.1"
description = "Autonomous Robot for General Onboard Surveillance -- ship inspection system"
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.11"
authors = [
    {name = "Fincantieri Digital", email = "digital@fincantieri.it"},
]
keywords = ["maritime", "inspection", "defect-detection", "onnx", "robotics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Manufacturing",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "onnxruntime>=1.17.0",
    "opencv-python>=4.9.0",
    "python-can>=4.3.0",
    "websockets>=12.0",
    "fastapi>=0.110.0",
    "pydantic-settings>=2.2.0",
    "pydantic>=2.6.0",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
    "uvicorn>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[tool.setuptools.packages.find]
include = ["argos*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
